requirements.txt
numpy~=1.24.2
pillow~=9.4.0
plyfile~=0.7.4
scipy~=1.10.1
matplotlib~=3.7.0